+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## Extra Icons Change Log
 
+### 0.2 (2018/08/25)
+* support Jenkins `jenkinsfile`, Gitlab `.gitlab-ci.yml` files.
+
 ### 0.1 (2018/08/23)
-* support Travis YML `*travis.yml` `*travis.yaml`, Appveyor YML `*appveyor.yml` `*appveyor.yaml` files.
+* support Travis `travis.yml`, Appveyor `appveyor.yml` files.